agent/hello-agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=2.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
